exercises/authorization/requirements.txt
flask==2.0.1
Werkzeug==2.0.1
requests==2.31.0
python-dotenv==1.0.0
oauthlib==3.2.2
pyOpenSSL==24.0.0
Flask-Login==0.6.3
requests_oauthlib==1.3.1